Got sac-discrete running with mathenv and transformer encoder model

+++ utilities/data_structures/Replay_Buffer.py
@@ -40,10 +40,12 @@ class Replay_Buffer(object):
             
     def separate_out_data_types(self, experiences):
         """Puts the sampled experience into the correct format for a PyTorch neural network"""
-        states = torch.from_numpy(np.vstack([e.state for e in experiences if e is not None])).float().to(self.device)
+        # TODO: Fix removed .float() from states
+        states = torch.from_numpy(np.vstack([e.state for e in experiences if e is not None])).to(self.device)
         actions = torch.from_numpy(np.vstack([e.action for e in experiences if e is not None])).float().to(self.device)
         rewards = torch.from_numpy(np.vstack([e.reward for e in experiences if e is not None])).float().to(self.device)
-        next_states = torch.from_numpy(np.vstack([e.next_state for e in experiences if e is not None])).float().to(self.device)
+        # TODO: Fix removed .float() from next_states
+        next_states = torch.from_numpy(np.vstack([e.next_state for e in experiences if e is not None])).to(self.device)
         dones = torch.from_numpy(np.vstack([int(e.done) for e in experiences if e is not None])).float().to(self.device)
         
         return states, actions, rewards, next_states, dones
